refactoring SceneXplain plugin
