+++ src/autogpt_plugins/scenex/__init__.py
@@ -1,8 +1,10 @@
 """This is a SceneX plugin for describing images for Auto-GPT."""
-from typing import Any, Dict, List, Optional, Tuple, TypedDict, TypeVar
-
+import os
+from typing import Any, Dict, List, Optional, Tuple, TypeVar, TypedDict
 from auto_gpt_plugin_template import AutoGPTPluginTemplate
 from colorama import Fore
+
+from .scenex_plugin import SceneXplain
 
 PromptGenerator = TypeVar("PromptGenerator")
 
@@ -19,21 +21,25 @@ class AutoGPTSceneXPlugin(AutoGPTPluginTemplate):
 
     def __init__(self):
         super().__init__()
-        self._name = "Auto-GPT-Scenex-Plugin"
+        self._name = "ImageExplainer"
         self._version = "0.0.1"
-        self._description = "Auto-GPT Scenex Plugin: Describe any image by URL."
+        self._description = (
+            "An Image Captioning Tool: Use this tool to generate a detailed caption for an image. "
+            "The input can be an image file of any format, and "
+            "the output will be a text description that covers every detail of the image."
+        )
+        self.scenexplain = SceneXplain(os.getenv("SCENEX_API_KEY"))
+        self._is_api_key_set = bool(os.getenv("SCENEX_API_KEY"))
 
     def post_prompt(self, prompt: PromptGenerator) -> PromptGenerator:
-        from .scenex_plugin.scenex_plugin import describe_image, is_api_key_set
-
-        if is_api_key_set():
+        if self._is_api_key_set:
             prompt.add_command(
-                "Describe image by URL",
+                self._description,
                 "describe_image",
                 {
                     "image": "<image>",
                 },
-                describe_image,
+                self.scenexplain.describe_image,
             )
         else:
             print(
